src/backend/csc_440_project_backend/requirements.txt
asn1crypto==0.24.0
cffi==1.12.3
cryptography==2.7
Django==2.2.4
django-cors-headers==3.1.0
django-rest-knox==4.1.0
djangorestframework==3.10.3
pip-autoremove==0.9.1
pycparser==2.19
pytz==2019.2
six==1.12.0
sqlparse==0.3.0